Added DTOs, Repos, Requests and Services. Only to transfers for now, which happen in AuthController, might need to move to own controller / rename some things

+++ database/migrations/2025_05_06_102515_create_transfers_table.php
@@ -14,8 +14,8 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('vehicle_id')->constrained('vehicles')->cascadeOnDelete();
             $table->boolean('materials_check')->default('false');
-            $table->integer('cash_before');
-            $table->integer('cash_after')->nullable();
+            $table->decimal('cash_before');
+            $table->decimal('cash_after')->nullable();
             $table->integer('km_start');
             $table->integer('km_end')->nullable();
             $table->boolean('is_done')->default('false');
